Refactor networkx graphs out of plot_nets
